+Grating stims, +Load movies, +movie_transformer
Added to full_pipeline.py .

Movie transformer must be in the same folder as the natural_movie folder in order to function and only needs to be run once.

Added regimen file to my repo - contains the parameters I was using when pkl error was produced. Use this to replicate.